# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OOP Demo API",
        "description": "API para demostrar conceptos de OOP en Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://tu-dominio-produccion.com",
            "description": "Servidor de Producción"
        },
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
        "/api/employees": {
            "get": {
                "tags": [
                    "Employee CRUD"
                ],
                "summary": "Obtener todos los empleados",
                "description": "Obtiene lista completa de empleados desde Supabase",
                "operationId": "9887e2f12f56dfc0e8c450ebfc4235c3",
                "responses": {
                    "200": {
                        "description": "Lista de empleados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Employee CRUD"
                ],
                "summary": "Crear nuevo empleado",
                "description": "Crea un empleado en Supabase con validaciones OOP",
                "operationId": "8615a8e8b4d85f9652265e929a1c5926",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana García"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 28
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana.garcia@empresa.com"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 75000
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Desarrolladora Senior"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empleado creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employee CRUD"
                ],
                "summary": "Obtener empleado por ID",
                "description": "Obtiene un empleado específico desde Supabase",
                "operationId": "a2d2d41265cbce32d06dbb314debd172",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Empleado no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Empleado no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Employee CRUD"
                ],
                "summary": "Actualizar empleado",
                "description": "Actualiza un empleado existente en Supabase",
                "operationId": "73decaec2a06938575bd21308c78b05f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ana García Actualizada"
                                    },
                                    "age": {
                                        "type": "integer",
                                        "example": 29
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "ana.garcia.nueva@empresa.com"
                                    },
                                    "salary": {
                                        "type": "number",
                                        "example": 80000
                                    },
                                    "position": {
                                        "type": "string",
                                        "example": "Tech Lead"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Empleado actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empleado actualizado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Employee CRUD"
                ],
                "summary": "Eliminar empleado",
                "description": "Elimina un empleado de Supabase",
                "operationId": "acde1b504fb1a5c7a555ac2f1b34daf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Empleado eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}